+++ Schedule/Controls/SSchedulePlaceholderView.xib
@@ -10,7 +10,7 @@
     <objects>
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" id="iN0-l3-epB" customClass="SNibSchedulePlaceholder" customModule="Schedule" customModuleProvider="target">
+        <view contentMode="scaleToFill" id="iN0-l3-epB" customClass="SNibSchedulePlaceholderView" customModule="Schedule" customModuleProvider="target">
             <rect key="frame" x="0.0" y="0.0" width="592" height="761"/>
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
             <subviews>
