+++ Schedule/Data/SCacheItem.swift
