+++ Schedule/Data/SClass.swift
