+++ Schedule/Data/SCourse.swift
