+++ Schedule/Data/SScheduleDay.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+final class SScheduleDay: Codable {
+    
+    let weekDay: SWeekDay
+    var classes: [SClass]
+    
+    init(weekDay: SWeekDay, classes: [SClass]) {
+        self.weekDay = weekDay
+        self.classes = classes
+    }
+    
+}
