+++ Schedule/en.lproj/InfoPlist.strings
@@ -1,3 +1,2 @@
 "CFBundleDisplayName" = "Schedule";
 "NSCalendarsUsageDescription" = "App needs your permission to access calendar for exporting schedule";
-"CFBundleName" = "Schedule";
